requirements.txt
# MegaETH Telegram Bot - Railway Deployment
# Python 3.11+

python-telegram-bot==21.0.1
aiohttp==3.9.3
aiosqlite==0.19.0
python-dateutil==2.8.2
